pyproject.toml
[tool.poetry]
name = "gemini"
version = "0.1.0"
description = "The server moderation bot for the Nextcord server."
authors = ["vcokltfre <vcokltfre@gmail.com>"]
repository = "https://github.com/nextcord/gemini"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
nextcord = "^2.0.0-alpha.1"
taskipy = "^1.8.1"
loguru = "^0.5.3"
python-dotenv = "^0.19.0"
asyncpg = "^0.24.0"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
isort = "^5.9.3"

[tool.taskipy.tasks]
start = "python3 -m src"
format = "black . && isort --profile black ."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
